Exclude com.codename1.impl from javadocs, add missing package-info, fix BX unboxing + CI gates
Three related documentation/quality fixes plus CI guards so they cannot
regress into master:

- javadoc: com.codename1.impl (and sub packages such as
  com.codename1.impl.gpu) was leaking into the published API. javadoc's
  `-exclude` only filters `-subpackages` discovery, which has no effect
  when sources are passed as an explicit @argfile. build_javadocs.sh now
  drops impl from the source list, adds `-sourcepath` so documented
  classes still resolve references to impl, and hard-fails if impl HTML
  is generated.

- package-info: add package-info.java to every documented package that
  lacked one (com.codename1.gpu, the box2d sub packages, plus camera,
  crash, annotations.graphql, annotations.rest). New
  scripts/check-package-info.sh fails the build when a non-impl package
  is missing its description; wired into the javadocs workflow.

- ByteCodeTranslator: fix the two BX_UNBOXING_IMMEDIATELY_REBOXED
  findings in JavascriptMethodGenerator (Integer cast to (int) only to
  be reboxed for a Map<Integer,Integer> lookup). Added
  BX_UNBOXING_IMMEDIATELY_REBOXED to the forbidden SpotBugs rules so a
  recurrence fails CI.

+++ vm/ByteCodeTranslator/src/com/codename1/tools/translator/JavascriptMethodGenerator.java
@@ -3824,7 +3824,7 @@ final class JavascriptMethodGenerator {
                 }
                 for (Label l : swLabels) {
                     Integer t = l == null ? null : labelToIndex.get(l);
-                    Integer tb = t == null ? null : startToBlock.get((int) t);
+                    Integer tb = t == null ? null : startToBlock.get(t);
                     if (tb == null) {
                         return _sb(method, instructions, "L3116");
                     }
@@ -5847,7 +5847,7 @@ private static void appendJsBodyMethod(StringBuilder out, ByteCodeClass cls, Byt
                 if (sw.getLabels() != null) {
                     for (Label l : sw.getLabels()) {
                         Integer d = l == null ? null : labelToIndex.get(l);
-                        if (d != null && startToBlock.containsKey((int) d)) { b.succs.add(startToBlock.get((int) d)); }
+                        if (d != null && startToBlock.containsKey(d)) { b.succs.add(startToBlock.get(d)); }
                     }
                 }
                 addFallThrough = false;
